Complete Custom Response
-- Student Dto , Entity , Service, repo creation
-- Student Save and get List api develop

+++ util/custom-response/src/main/java/com/brother/customresponse/payload/Response.java
@@ -21,7 +21,7 @@ public class Response {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private Object content;
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    private int numberOfElement;
+    private Long numberOfElement;
     @JsonInclude(JsonInclude.Include.NON_NULL)
     private Long rowCount;
     @JsonInclude(JsonInclude.Include.NON_NULL)
